Checkpoint: bugfixes and more setup 2
